Work-around for the TF2Items_OnGiveNamedItem bug
Instead of replacing weapons in the forward, this patch
creates a timer to equip the correct weapon 0.3s later
using TF2Items_GiveNamedItem.
If you want to disable this behaviour and use the old
OnGiveNamedItem method for some reason, comment out
this line at the top of the file #define USE_GIVEITEM_TIMER
and recompile the plugin.