Fixed double chat output
This is based on the wild assumption that the messages show up
twice if a player is fully connected and only once on join.
And by twice i mean ALWAYS twice and are not racy. hahaha.
This might need to be fixed, because currently it just
toggles a bool whether to show the next line or not - and as
soon as they show up out of order e.g. ABAB it would show AA.
On the other hand this is not as bad as it might sound, because
players can only change one weapon at a time. Except when they
join, they equip all weapons at once - but we've catched that
with the IgnoreNextLine toggle not being triggered then.
So I guess this could really work... needs testing though.
Oh btw, the toggle is player specific, so it does not matter
whether to players switch their weapons at the same time.