Fixed values being overwritten by prefabs
This resulted in item names being overwritten
by the value in the prefab and hence showing e.g.
'Rocket Launcher' for all Botkiller and Festive
Weapons, instead of 'Festive Rocket Launcher'.
This seems to be a cosmetic fix, but in reality
isnt, cause not only the items name was affected.
There are no other issues known that were caused
by this though.
Odd.